+++ pkg/controller/volume/persistentvolume/binder_test.go
@@ -22,7 +22,7 @@ import (
 	v1 "k8s.io/api/core/v1"
 	storage "k8s.io/api/storage/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	pvutil "k8s.io/kubernetes/pkg/controller/volume/persistentvolume/util"
+	pvutil "k8s.io/component-helpers/storage/volume"
 )
 
 // Test single call to syncClaim and syncVolume methods.
