+++ pkg/controller/volume/persistentvolume/delete_test.go
@@ -20,10 +20,10 @@ import (
 	"errors"
 	"testing"
 
-	"k8s.io/api/core/v1"
+	v1 "k8s.io/api/core/v1"
 	storage "k8s.io/api/storage/v1"
+	pvutil "k8s.io/component-helpers/storage/volume"
 	pvtesting "k8s.io/kubernetes/pkg/controller/volume/persistentvolume/testing"
-	pvutil "k8s.io/kubernetes/pkg/controller/volume/persistentvolume/util"
 )
 
 // Test single call to syncVolume, expecting recycling to happen.
