+++ pkg/controller/volume/persistentvolume/pv_controller_test.go
@@ -36,21 +36,14 @@ import (
 	"k8s.io/client-go/tools/cache"
 	"k8s.io/component-base/featuregate"
 	featuregatetesting "k8s.io/component-base/featuregate/testing"
+	pvutil "k8s.io/component-helpers/storage/volume"
 	csitrans "k8s.io/csi-translation-lib"
 	"k8s.io/klog/v2"
 	"k8s.io/kubernetes/pkg/controller"
 	pvtesting "k8s.io/kubernetes/pkg/controller/volume/persistentvolume/testing"
-	pvutil "k8s.io/kubernetes/pkg/controller/volume/persistentvolume/util"
 	"k8s.io/kubernetes/pkg/features"
 	"k8s.io/kubernetes/pkg/volume/csimigration"
 	"k8s.io/kubernetes/pkg/volume/util"
-)
-
-var (
-	classNotHere       = "not-here"
-	classNoMode        = "no-mode"
-	classImmediateMode = "immediate-mode"
-	classWaitMode      = "wait-mode"
 )
 
 // Test the real controller methods (add/update/delete claim/volume) with
@@ -469,88 +462,12 @@ func TestControllerCacheParsingError(t *testing.T) {
 	}
 }
 
-func makePVCClass(scName *string) *v1.PersistentVolumeClaim {
-	claim := &v1.PersistentVolumeClaim{
-		ObjectMeta: metav1.ObjectMeta{
-			Annotations: map[string]string{},
-		},
-		Spec: v1.PersistentVolumeClaimSpec{
-			StorageClassName: scName,
-		},
-	}
-
-	return claim
-}
-
 func makeStorageClass(scName string, mode *storagev1.VolumeBindingMode) *storagev1.StorageClass {
 	return &storagev1.StorageClass{
 		ObjectMeta: metav1.ObjectMeta{
 			Name: scName,
 		},
 		VolumeBindingMode: mode,
-	}
-}
-
-func TestDelayBindingMode(t *testing.T) {
-	tests := map[string]struct {
-		pvc         *v1.PersistentVolumeClaim
-		shouldDelay bool
-		shouldFail  bool
-	}{
-		"nil-class": {
-			pvc:         makePVCClass(nil),
-			shouldDelay: false,
-		},
-		"class-not-found": {
-			pvc:         makePVCClass(&classNotHere),
-			shouldDelay: false,
-		},
-		"no-mode-class": {
-			pvc:         makePVCClass(&classNoMode),
-			shouldDelay: false,
-			shouldFail:  true,
-		},
-		"immediate-mode-class": {
-			pvc:         makePVCClass(&classImmediateMode),
-			shouldDelay: false,
-		},
-		"wait-mode-class": {
-			pvc:         makePVCClass(&classWaitMode),
-			shouldDelay: true,
-		},
-	}
-
-	classes := []*storagev1.StorageClass{
-		makeStorageClass(classNoMode, nil),
-		makeStorageClass(classImmediateMode, &modeImmediate),
-		makeStorageClass(classWaitMode, &modeWait),
-	}
-
-	client := &fake.Clientset{}
-	informerFactory := informers.NewSharedInformerFactory(client, controller.NoResyncPeriodFunc())
-	classInformer := informerFactory.Storage().V1().StorageClasses()
-	ctrl := &PersistentVolumeController{
-		classLister: classInformer.Lister(),
-		translator:  csitrans.New(),
-	}
-
-	for _, class := range classes {
-		if err := classInformer.Informer().GetIndexer().Add(class); err != nil {
-			t.Fatalf("Failed to add storage class %q: %v", class.Name, err)
-		}
-	}
-
-	for name, test := range tests {
-		shouldDelay, err := pvutil.IsDelayBindingMode(test.pvc, ctrl.classLister)
-		if err != nil && !test.shouldFail {
-			t.Errorf("Test %q returned error: %v", name, err)
-		}
-		if err == nil && test.shouldFail {
-			t.Errorf("Test %q returned success, expected error", name)
-		}
-		if shouldDelay != test.shouldDelay {
-			t.Errorf("Test %q returned unexpected %v", name, test.shouldDelay)
-		}
 	}
 }
